+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added new endpoint for multiple items packing for multiple boxes
+
 ## [0.3.0] - 2022-05-05
 
 ### Added